requirements.txt
mistralai
httpx
python-dotenv